--- src/zilean/__init__.py
@@ -1,2 +1,0 @@
-from .core import *
-from .TimelineCrawler import *
